+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.8.0] - 2026-03-16
+
 ### Changed
 
 - Update Edge driver versions map.
@@ -222,7 +224,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial version release.
 
-[Unreleased]: https://github.com/atata-framework/atata-webdriversetup/compare/v3.7.0...HEAD
+[Unreleased]: https://github.com/atata-framework/atata-webdriversetup/compare/v3.8.0...HEAD
+[3.8.0]: https://github.com/atata-framework/atata-webdriversetup/compare/v3.7.0...v3.8.0
 [3.7.0]: https://github.com/atata-framework/atata-webdriversetup/compare/v3.6.0...v3.7.0
 [3.6.0]: https://github.com/atata-framework/atata-webdriversetup/compare/v3.5.0...v3.6.0
 [3.5.0]: https://github.com/atata-framework/atata-webdriversetup/compare/v3.4.0...v3.5.0